Create useFormikContextWithOverrides hook
